moving keystore to foundation
