+++ business/auth/auth_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/ardanlabs/service/business/auth"
-	"github.com/ardanlabs/service/business/keystore"
+	"github.com/ardanlabs/service/foundation/keystore"
 	"github.com/dgrijalva/jwt-go/v4"
 )
